requirements.txt
Pillow==8.2.0
pyzbar==0.1.8
